refactor: move AmpersandProvider to context folder (#264)
- add folders to Configure component
- move NavObjectItem, OtherTab, UninstallInstallation to nav folder
- add constant file to nav folder
- move OAuthPopup to Oauth folder
- move AmpersandProvider to context folder
- update imports
