+++ src/components/Configure/nav/ObjectManagementNav/constant.tsx
@@ -1,0 +1,1 @@
+export const OTHER_CONST = 'other';
