+++ src/components/Oauth/NoSubdomainEntry/NoSubdomainOauthFlow.tsx
@@ -7,8 +7,8 @@ import { useCallback, useState } from 'react';
 import { useApiKey } from '../../../context/ApiKeyContextProvider';
 import { useProject } from '../../../context/ProjectContextProvider';
 import { capitalize } from '../../../utils';
-import OAuthPopup from '../../Connect/OAuthPopup';
 import { fetchOAuthCallbackURL } from '../fetchOAuthCallbackURL';
+import OAuthPopup from '../OAuthPopup';
 
 import { LandingContent } from './LandingContent';
