+++ src/components/Oauth/Salesforce/SalesforceOauthFlow.tsx
@@ -8,8 +8,8 @@ import { useCallback, useState } from 'react';
 import { PROVIDER_SALESFORCE } from '../../../constants';
 import { useApiKey } from '../../../context/ApiKeyContextProvider';
 import { useProject } from '../../../context/ProjectContextProvider';
-import OAuthPopup from '../../Connect/OAuthPopup';
 import { fetchOAuthCallbackURL } from '../fetchOAuthCallbackURL';
+import OAuthPopup from '../OAuthPopup';
 
 import { SubdomainEntry } from './SubdomainEntry';
